requirements.txt
discord.py
jishaku
tinydb
dict-deep